move cargohold endpoints to Public module, too
